add Metrics

+++ src/Chaldea.Fate.RhoAias/ClientManager.cs
@@ -1,34 +1,33 @@
 ﻿namespace Chaldea.Fate.RhoAias;
 
-public interface IServerManager
+public interface IClientManager
 {
 	Task<bool> RegisterClientAsync(Client register);
 	Task UnRegisterClientAsync(string connectionId);
 	Task CreateClientAsync(Client entity);
 	Task RemoveClientAsync(Guid id);
 	Task<List<Client>> GetClientListAsync();
-	Task CreateProxyAsync(Proxy entity);
-	Task RemoveProxyAsync(Guid id);
-	Task<List<Proxy>> GetProxyListAsync();
+	Task InitClientMetricsAsync();
+	Task ResetClientStatusAsync();
 }
 
-internal class ServerManager : IServerManager
+internal class ClientManager : IClientManager
 {
 	private readonly IRepository<Client> _clientRepository;
-	private readonly IRepository<Proxy> _proxyRepository;
 	private readonly IForwarderManager _forwarderManager;
 	private readonly ITokenManager _tokenManager;
+	private readonly IMetrics _metrics;
 
-	public ServerManager(
+	public ClientManager(
 		IRepository<Client> clientRepository,
-		IRepository<Proxy> proxyRepository,
 		IForwarderManager forwarderManager,
-		ITokenManager tokenManager)
+		ITokenManager tokenManager,
+		IMetrics metrics)
 	{
 		_clientRepository = clientRepository;
-		_proxyRepository = proxyRepository;
 		_forwarderManager = forwarderManager;
 		_tokenManager = tokenManager;
+		_metrics = metrics;
 	}
 
 	public async Task<bool> RegisterClientAsync(Client register)
@@ -45,12 +44,11 @@ internal class ServerManager : IServerManager
 			client.Proxies?.Clear();
 			client.Proxies = register.Proxies;
 		}
+
 		// register all proxies
-		if (client.Proxies is { Count: > 0 })
-		{
-			_forwarderManager.Register(client.Proxies.ToList());
-		}
+		if (client.Proxies is { Count: > 0 }) _forwarderManager.Register(client.Proxies.ToList());
 		await _clientRepository.UpdateAsync(client);
+		_metrics.UpDownClientOnline(1);
 		return true;
 	}
 
@@ -62,29 +60,26 @@ internal class ServerManager : IServerManager
 		// update status
 		client.Status = false;
 		// unregister proxies
-		if (client.Proxies is { Count: > 0 })
-		{
-			_forwarderManager.UnRegister(client.Proxies.ToList());
-		}
+		if (client.Proxies is { Count: > 0 }) _forwarderManager.UnRegister(client.Proxies.ToList());
 		await _clientRepository.UpdateAsync(client);
+		_metrics.UpDownClientOnline(-1);
 	}
 
 	public async Task CreateClientAsync(Client entity)
 	{
 		var cid = Guid.NewGuid();
 		entity.Id = cid;
-		entity.Token = await _tokenManager.CreateAsync(cid);
+		entity.Token = await _tokenManager.CreateAsync(cid, Role.Client, DateTime.UtcNow.AddYears(10));
 		await _clientRepository.InsertAsync(entity);
+		_metrics.UpDownClientTotal(1);
 	}
 
 	public async Task RemoveClientAsync(Guid id)
 	{
 		// Cascade Delete
 		var entity = await _clientRepository.GetAsync(x => x.Id == id, x => x.Proxies);
-		if (entity != null)
-		{
-			await _clientRepository.DeleteAsync(entity);
-		}
+		if (entity != null) await _clientRepository.DeleteAsync(entity);
+		_metrics.UpDownClientTotal(-1);
 	}
 
 	public async Task<List<Client>> GetClientListAsync()
@@ -93,30 +88,19 @@ internal class ServerManager : IServerManager
 		return clients;
 	}
 
-	public async Task CreateProxyAsync(Proxy entity)
+	public async Task InitClientMetricsAsync()
 	{
-		entity.Id = Guid.NewGuid();
-		await _proxyRepository.InsertAsync(entity);
-		var proxies = await _proxyRepository.GetListAsync(x => x.Id == entity.Id && x.Client.Status, x => x.Client);
-		if (proxies is { Count: > 0 })
-		{
-			_forwarderManager.Register(proxies);
-		}
+		var total = await _clientRepository.CountAsync();
+		_metrics.UpDownClientTotal(total);
 	}
 
-	public async Task RemoveProxyAsync(Guid id)
+	public async Task ResetClientStatusAsync()
 	{
-		var proxies = await _proxyRepository.GetListAsync(x => x.Id == id);
-		if (proxies is { Count: > 0 })
+		var clients = await _clientRepository.GetListAsync(x => x.Status);
+		foreach (var client in clients)
 		{
-			_forwarderManager.UnRegister(proxies);
-			await _proxyRepository.DeleteManyAsync(proxies);
+			client.Status = false;
 		}
-	}
-
-	public async Task<List<Proxy>> GetProxyListAsync()
-	{
-		var list = await _proxyRepository.GetListAsync(x => x.Client);
-		return list;
+		await _clientRepository.UpdateManyAsync(clients);
 	}
 }
